requirements.txt
pandas
numpy
scikit-learn
nltk
Pillow
tkinter
